Fix static analysis issues
:Release Notes:
Fix static analysis issues

:Detailed Notes:
Fix static analysis issues - Low

:Testing Performed:
UnitTest Passed

:QA Notes:
N/A

:Issues Addressed:
[WRR-4876] Fix static analysis issues

+++ headers/webosi18n/cxx/ResBundle.h
@@ -136,7 +136,7 @@ class ResBundle {
 		/**
 		 * Initialize pseudo maps
 		 */
-		map< int, map<string, string> > initPseudoCharMap();
+		map< size_t, map<string, string> > initPseudoCharMap();
 
 	private:
 		/**
@@ -154,7 +154,7 @@ class ResBundle {
 		/**
 		 * Char maps for pseudo localization
 		 */
-		map< int, map<string, string> > pseudomap;
+		map< size_t, map<string, string> > pseudomap;
 };
 
 #endif  /* RESBUNDLE_H_ */
